src/test/java/ru/skypro/homework/service/UserServiceTest.java
package ru.skypro.homework.service;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.security.core.Authentication;
import ru.skypro.homework.dto.UserDTO;
import ru.skypro.homework.mapper.UserMapper;
import ru.skypro.homework.model.User;
import ru.skypro.homework.repository.UserRepository;

import java.util.Optional;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNotNull;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.*;

@ExtendWith(MockitoExtension.class)
public class UserServiceTest {

    @Mock
    private UserRepository userRepository;


    @InjectMocks
    private UserService userService;

    @Test
    public void testGetAuthorizedUsers() throws Exception {
        String username = "testuser";
        User user = new User();
        user.setUsername(username);

        when(userRepository.findByUsernameIgnoreCase(username)).thenReturn(Optional.of(user));

        Authentication authentication = mock(Authentication.class);
        when(authentication.getName()).thenReturn(username);


        UserDTO result = userService.getAuthorizedUsers(authentication);

        assertNotNull(result);
        assertEquals(username, result.getEmail());
        verify(userRepository).findByUsernameIgnoreCase(username);
    }

}